% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevodata-getters.R
\name{active_models}
\alias{active_models}
\title{Get the name of the active models}
\usage{
active_models(object)
}
\arguments{
\item{object}{cebodata object}
}
\description{
Get the name of the active models
}
