% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevo_snvs.R
\name{annotate_mutation_contexts}
\alias{annotate_mutation_contexts}
\title{Annotate mutation context and types for mutation signatures analysis}
\usage{
annotate_mutation_contexts(snvs, bsgenome)
}
\arguments{
\item{snvs}{snvs tbl}

\item{bsgenome}{BSGenome object}
}
\description{
Annotate mutation context and types for mutation signatures analysis
}
