% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevo_cnvs.R
\name{annotate_normal_cn}
\alias{annotate_normal_cn}
\title{Annotate chromosome ploidies in CNV data}
\usage{
annotate_normal_cn(object, which_cnvs = default_CNVs(object))
}
\arguments{
\item{object}{\if{html}{\out{<cevodata>}} object}

\item{which_cnvs}{Name of the CNVs slot}
}
\value{
\if{html}{\out{<cevodata>}} object
}
\description{
Adds the normal_cn column to the data. This column is required for e.g.
by Dentro CCF calculation method. Requires 'sex' column in the metadata.
Males should be encoded by 'M' or "male'.
}
