% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevodata-class.R, R/generics.R
\name{SNVs.cevodata}
\alias{SNVs.cevodata}
\alias{add_SNV_data.cevodata}
\alias{CNVs.cevodata}
\alias{add_CNV_data.cevodata}
\alias{assays}
\alias{SNVs}
\alias{add_SNV_data}
\alias{CNVs}
\alias{add_CNV_data}
\title{Get/Add SNV/CNV data from the cevodata dataset}
\usage{
\method{SNVs}{cevodata}(object, which = object$active_SNVs, ...)

\method{add_SNV_data}{cevodata}(object, snvs, name = NULL, ...)

\method{CNVs}{cevodata}(object, which = object$active_CNVs, ...)

\method{add_CNV_data}{cevodata}(object, cnvs, name = NULL, ...)

SNVs(object, ...)

add_SNV_data(object, ...)

CNVs(object, ...)

add_CNV_data(object, ...)
}
\arguments{
\item{object}{object}

\item{which}{assay to use - uses active_SNVs if nonei}

\item{...}{other arguments}

\item{snvs}{tibble with SNVs}

\item{name}{name for SNVs/CNVs assay}

\item{cnvs}{tibble with CNVs}
}
\value{
tibble
}
\description{
Get/Add SNV/CNV data from the cevodata dataset
}
\section{Functions}{
\itemize{
\item \code{SNVs(cevodata)}: Get SNVs from cevodata dataset

\item \code{add_SNV_data(cevodata)}: Add new SNVs to cevodata

\item \code{CNVs(cevodata)}: Get CNVs from cevodata dataset

\item \code{add_CNV_data(cevodata)}: Add new CNVs to cevodata

}}
