% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_subclones_binomial_CliP.R
\name{build_CliP_container}
\alias{build_CliP_container}
\title{Build CliP Apptainer container}
\usage{
build_CliP_container(out_dir = ".")
}
\arguments{
\item{out_dir}{path to save the CliP.sif file}
}
\description{
Build CliP Apptainer container
}
