% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_tails.R
\name{calc_cumulative_tails}
\alias{calc_cumulative_tails}
\title{Calculate the cumulative tails}
\usage{
calc_cumulative_tails(dt, digits = 2)
}
\arguments{
\item{dt}{tibble with VAF column. Might be groupped and contain the data for multiple samples}

\item{digits}{resolution of the cumulative tails calculation}
}
\value{
tbl with the groupping variables and:
\itemize{
\item n columnt with the number of mutations in the VAF interval
\item x and y columns describing the cumulative tails
\item y_scaled with y values scaled to the range 0-1
}
}
\description{
Calculate the cumulative tails
}
\examples{
data("snvs_test")
snvs_test |>
  dplyr::group_by(sample_id) |>
  calc_cumulative_tails()
}
