% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/williams_Mf_1f.R
\name{calc_Mf_1f}
\alias{calc_Mf_1f}
\title{Calculate Williams M(f) ~ 1/f}
\usage{
calc_Mf_1f(dt, digits = 2)
}
\arguments{
\item{dt}{tibble with VAF column. Might be groupped and contain the data for multiple samples}

\item{digits}{resolution of the cumulative tails calculation}
}
\value{
tbl with the groupping variables and:
\itemize{
\item VAF
\item n - number of mutations in the VAF interval
\item \code{M(f)} and \code{1/f} columns to plot William's statistics
}
}
\description{
Calculate Williams M(f) ~ 1/f
}
\examples{
data("tcga_brca_test")
tcga_brca_test |>
  dplyr::group_by(sample_id) |>
  calc_Mf_1f()
}
