% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_snvs.R
\name{calc_mutation_frequencies}
\alias{calc_mutation_frequencies}
\alias{dentro_2015_correction}
\title{Calc mutation frequencies}
\usage{
calc_mutation_frequencies(
  object,
  method = "Dentro",
  which_snvs = default_SNVs(object),
  which_cnvs = default_CNVs(object),
  rm_intermediate_cols = TRUE,
  verbose = get_cevomod_verbosity()
)

dentro_2015_correction(tbl)
}
\arguments{
\item{object}{\if{html}{\out{<cevodata>}} object}

\item{method}{Available methods: Dentro}

\item{which_snvs}{Which SNVs to use}

\item{which_cnvs}{Which CNVs to use}

\item{rm_intermediate_cols}{Should the columns used to get CCF be removed?}

\item{verbose}{Verbose?}

\item{tbl}{tibble that contains columns: VAF, total_cn, normal_cn, purity}
}
\value{
\if{html}{\out{<cevodata>}} object

The same tibble with new columns: u, m, CCF (Cancer Cell Fraction)
}
\description{
Calculates the CNV-corrected mutation frequencies. Implemented methods:
\itemize{
\item Dentro - calculates the Cancer Cell Fraction (CCF) using the formulas from
\href{https://doi.org/10.1101/cshperspect.a026625}{Dentro et al. \emph{Principles of Reconstructing the Subclonal Architecture of Cancers} (2015)}
}
}
\section{Functions}{
\itemize{
\item \code{dentro_2015_correction()}: Implements the CNV-based frequency
correction method described in \href{https://doi.org/10.1101/cshperspect.a026625}{Dentro et al. 'Principles of Reconstructing the Subclonal Architecture of Cancers' (2015)}

}}
