% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFS.R
\name{calc_SFS}
\alias{calc_SFS}
\title{Calculate SFS}
\usage{
calc_SFS(dt, digits = 2)
}
\arguments{
\item{dt}{tibble with VAF column. Might be groupped and contain the data for multiple samples}

\item{digits}{resolution of the cumulative tails calculation}
}
\value{
tbl with the groupping variables and:
\itemize{
\item n columnt with the number of mutations in the VAF interval
\item x and y columns describing SFS
\item y_scaled with y values scaled to the range 0-1
}
}
\description{
Calculate SFS
}
\examples{
data("tcga_brca_test")
tcga_brca_test |>
  dplyr::group_by(sample_id) |>
  calc_SFS()
}
