% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV_plots.R
\name{cnv_plots}
\alias{cnv_plots}
\alias{plot_CNV_heatmap}
\alias{plot_CNV_heatmap.cevodata}
\alias{heatmap_granges}
\title{Plot CNV heatmap}
\usage{
plot_CNV_heatmap(object, ...)

\method{plot_CNV_heatmap}{cevodata}(object, meta_field, ...)

heatmap_granges(
  granges,
  meta_field,
  row_groups = NULL,
  keep_sites_present_in = floor(0.8 * length(granges)),
  color_breaks = c(0, 2, 6),
  colors = c("dodgerblue3", "white", "firebrick3"),
  window_width = 1e+06,
  upper_limit = 6,
  cluster_rows = FALSE,
  show_row_names = TRUE,
  show_column_names = FALSE,
  use_raster = TRUE,
  cluster_columns = FALSE,
  border = TRUE,
  legend_params = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{cevodata}

\item{...}{other arguments passed to Heatmap}

\item{meta_field}{column to plot}

\item{granges}{granges list}

\item{row_groups}{vector}

\item{keep_sites_present_in}{floor(0.8*length(granges))}

\item{color_breaks}{c(0, 2, 6)}

\item{colors}{c("dodgerblue3", "white", "firebrick3")}

\item{window_width}{window_width}

\item{upper_limit}{upper limit of color scale}

\item{cluster_rows}{FALSE}

\item{show_row_names}{TRUE}

\item{show_column_names}{FALSE}

\item{use_raster}{TRUE}

\item{cluster_columns}{FALSE}

\item{border}{TRUE}

\item{legend_params}{goes to Heatmap(heatmap_legend_param = .)}

\item{verbose}{Verbose?}
}
\description{
Plot CNV heatmap
}
\section{Functions}{
\itemize{
\item \code{plot_CNV_heatmap(cevodata)}: Plot CNVs

\item \code{heatmap_granges()}: Plot Granges list

}}
