% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_and_themes.R
\docType{data}
\name{colors}
\alias{colors}
\title{Some colors}
\format{
An object of class \code{cevo_colors} (inherits from \code{list}) of length 2.
}
\usage{
colors
}
\description{
Some colors
}
\keyword{datasets}
