% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_models.R
\name{compare_models}
\alias{compare_models}
\title{Compare fits from many models}
\usage{
compare_models(object, model_names, column_name, ...)
}
\arguments{
\item{object}{cevodata object}

\item{model_names}{models to compare}

\item{column_name}{residuals_* column to plot}

\item{...}{other arguments passed to plot_SFS()}
}
\description{
Compare fits from many models
}
