% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevo_snvs.R
\name{count_mutation_types}
\alias{count_mutation_types}
\title{Get matrix of per sample mutation types for mutation signatures analysis}
\usage{
count_mutation_types(snvs)
}
\arguments{
\item{snvs}{annotated snvs tbl}
}
\value{
wide tbl
}
\description{
Get matrix of per sample mutation types for mutation signatures analysis
}
