% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{count_mutations_by_component}
\alias{count_mutations_by_component}
\alias{count_neutral_tail_mutations}
\alias{count_neutral_tail_filtered_mutations}
\title{Get numbers of neutral, clonal, subclonal ect variants}
\usage{
count_mutations_by_component(
  object,
  models_name = active_models(object),
  include_filtered = FALSE
)

count_neutral_tail_mutations(object, models_name = active_models(object))

count_neutral_tail_filtered_mutations(
  object,
  models_name = active_models(object)
)
}
\arguments{
\item{object}{cevodata object}

\item{models_name}{models name}

\item{include_filtered}{include the component of the filtered mutations}
}
\description{
Get numbers of neutral, clonal, subclonal ect variants
}
\section{Functions}{
\itemize{
\item \code{count_neutral_tail_mutations()}: Count neutral tail mutations

\item \code{count_neutral_tail_filtered_mutations()}: Count filtered mutations with VAF higher than 0.01

}}
