% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{cut_f_intervals}
\alias{cut_f_intervals}
\title{Get mutation frequency intervals}
\usage{
cut_f_intervals(snvs, column, bins = NULL)
}
\arguments{
\item{snvs}{SNVs tibble}

\item{column}{Name of column with frequencies, eg. VAF or CCF}

\item{bins}{Number of bins}
}
\value{
Tibble with f_interval and f columns
}
\description{
Cuts requested column e.g. CCF or VAF into intervals. Adds f_interval and f
columns to the input SNVs tibble
}
