% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{cut_VAF_intervals}
\alias{cut_VAF_intervals}
\alias{cut_VAF_intervals.cevo_snvs}
\title{Cut VAF into intervals}
\usage{
cut_VAF_intervals(object, ...)

\method{cut_VAF_intervals}{cevo_snvs}(object, bins = NULL, ...)
}
\arguments{
\item{object}{object}

\item{...}{other params}

\item{bins}{number of bins}
}
\description{
Cut VAF into intervals
}
