% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevomod.R
\docType{data}
\name{driver_genes}
\alias{driver_genes}
\title{Tbl of driver genes from Bailey, Ding et al. 'Comprehensive Characterization
of Cancer Driver Genes and Mutations', Cell, 2-18
https://doi.org/10.1016/j.cell.2018.02.060}
\description{
Tbl of driver genes from Bailey, Ding et al. 'Comprehensive Characterization
of Cancer Driver Genes and Mutations', Cell, 2-18
https://doi.org/10.1016/j.cell.2018.02.060
}
