% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/powerlaw_neutral_tail.R
\name{estimate_sampling_rate}
\alias{estimate_sampling_rate}
\alias{estimate_sampling_rate.cevodata}
\title{estimate_sampling_rate}
\usage{
estimate_sampling_rate(object, ...)

\method{estimate_sampling_rate}{cevodata}(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{other arguments}
}
\value{
tibble
}
\description{
estimate_sampling_rate
}
\section{Methods (by class)}{
\itemize{
\item \code{estimate_sampling_rate(cevodata)}: Uses experimental SFS and power-law model
to estimate the sampling rate

}}
