% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_joining.R
\name{evaluate_MC_runs}
\alias{evaluate_MC_runs}
\title{Evaluate Monte Carlo results}
\usage{
evaluate_MC_runs(mc_arr, rowsums_pred, colsums_pred)
}
\arguments{
\item{mc_arr}{Monte Carlo simutations array}

\item{rowsums_pred}{expected row sums}

\item{colsums_pred}{predicted col sums}
}
\description{
Evaluate Monte Carlo results
}
