% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fill_na}
\alias{fill_na}
\title{Fill na values in the object}
\usage{
fill_na(object, val)
}
\arguments{
\item{object}{object}

\item{val}{value to fill the NAs}
}
\description{
Fill na values in the object
}
