% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevodata-transformations.R
\name{filter.cevodata}
\alias{filter.cevodata}
\title{Filter/subset cevodata object}
\usage{
\method{filter}{cevodata}(.data, ..., .preserve = FALSE)
}
\arguments{
\item{.data}{cevodata object}

\item{...}{expression passed to dplyr::filter(...)}

\item{.preserve}{Relevant when the \code{.data} input is grouped.
If \code{.preserve = FALSE} (the default), the grouping structure
is recalculated based on the resulting data, otherwise the grouping is kept as is.}
}
\value{
cevodata object
}
\description{
This is a wrapper around dplyr::filter function which can be used to subset
cevodata object. Works like dplyr::filter, performs the filtering on metadata,
then filters SNVs, CNVs, clones and models keeping samples kept in metadata
}
