% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevo_snvs.R
\name{filter_SNVs_by_regions}
\alias{filter_SNVs_by_regions}
\title{Filter SNVs by position: using regions tbl or bed file}
\usage{
filter_SNVs_by_regions(snvs, regions = NULL, bed_file = NULL)
}
\arguments{
\item{snvs}{snvs tbl with columns: sample_id, chrom, pos}

\item{regions}{regions tbl with columns chrom, start, end}

\item{bed_file}{bed file}
}
\description{
Filter SNVs by position: using regions tbl or bed file
}
