% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{fit_mobster}
\alias{fit_mobster}
\alias{fit_mobster.cevodata}
\title{Fit Mobster models}
\usage{
fit_mobster(object, ...)

\method{fit_mobster}{cevodata}(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{other args passed to mobster::fit_mobster()}
}
\description{
Fit Mobster models
}
