% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutral_lm.R
\name{fit_neutral_lm}
\alias{fit_neutral_lm}
\title{Fit Williams neutral models to the data}
\usage{
fit_neutral_lm(dt, rsq_treshold = 0.98)
}
\arguments{
\item{dt}{tibble with results from calc_Mf_1f()}

\item{rsq_treshold}{R-squared tresholds to keep model as neutral}
}
\value{
cevo_lm_models_tbl
}
\description{
Fit Williams neutral models to the data
}
\examples{
Mf_1f <- tcga_brca_test |>
  dplyr::filter(sample_id \%in\% c("TCGA-AC-A23H-01","TCGA-AN-A046-01")) |>
  dplyr::group_by(sample_id) |>
  calc_Mf_1f()
models <- fit_neutral_lm(Mf_1f, rsq_treshold = 0.99)
plot(Mf_1f, from = 0.05, to = 0.4, scale = FALSE) +
  layer_lm_fits(models)
}
