% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevodata-getters.R
\name{fix_powerlaw_N_mutations}
\alias{fix_powerlaw_N_mutations}
\title{Fill N_mutations column for powerlaw models}
\usage{
fix_powerlaw_N_mutations(models, cd, models_name)
}
\arguments{
\item{models}{tibble from get_models()}

\item{cd}{cevodata object}

\item{models_name}{models name}
}
\description{
Fill N_mutations column for powerlaw models
}
