% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generators.R
\name{generate_neutral_snvs}
\alias{generate_neutral_snvs}
\title{Generate neutral SNVs}
\usage{
generate_neutral_snvs(mut_rate = 2, sample_below = 0.15, resolution = 0.01)
}
\arguments{
\item{mut_rate}{a in f(x) = a/x^2 formula}

\item{sample_below}{mutations below this value will be sampled}

\item{resolution}{resolution}
}
\value{
SNVs tibble
}
\description{
Generate neutral SNVs
}
\examples{
generate_neutral_snvs()
}
