% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_models.R
\name{geom_powerlaw}
\alias{geom_powerlaw}
\title{Plot power-law curve}
\usage{
geom_powerlaw(A, alpha, mapping, ylim = 1000, color = "#54b4FA", ...)
}
\arguments{
\item{A}{A}

\item{alpha}{power}

\item{mapping}{mapping, x is required}

\item{ylim}{max y allowed}

\item{color}{color}
}
\description{
Plot power-law curve
}
