% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolutionary_parameters.R
\name{get_evolutionary_parameters}
\alias{get_evolutionary_parameters}
\title{Get evolutionary parameters}
\usage{
get_evolutionary_parameters(object)
}
\description{
Get evolutionary parameters
}
