% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_snvs.R
\name{get_frequency_measure_name}
\alias{get_frequency_measure_name}
\title{Get mutation rates by Williams
Use ...}
\usage{
get_frequency_measure_name(object, ...)
}
\description{
Get mutation rates by Williams
Use ...
}
