% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_powerlaw_tail_fixed.R
\name{get_model_coefficients}
\alias{get_model_coefficients}
\title{Get model coefficients}
\usage{
get_model_coefficients(
  object,
  model_name = active_models(object),
  best_only = TRUE
)
}
\arguments{
\item{object}{cevodata object}

\item{model_name}{Model name}

\item{best_only}{Return only best models?}
}
\description{
Get model coefficients
}
