% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_powerlaw_tail_fixed.R
\name{get_model_residuals}
\alias{get_model_residuals}
\title{Get model residuals}
\usage{
get_model_residuals(object, model_name = active_models(object))
}
\arguments{
\item{object}{cevodata object}

\item{model_name}{Model name}
}
\description{
Get model residuals
}
