% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevodata-construction.R
\name{get_models}
\alias{get_models}
\alias{get_models.cevodata}
\title{Get models from the object}
\usage{
get_models(object, ...)

\method{get_models}{cevodata}(object, which = active_models(object), ...)
}
\arguments{
\item{which}{\code{chr} which models to get}
}
\description{
Get models from the object
}
\section{Methods (by class)}{
\itemize{
\item \code{get_models(cevodata)}: Get models from cevodata object

}}
