% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolutionary_parameters.R
\name{get_mutation_rates}
\alias{get_mutation_rates}
\title{Get mutation rates by Williams}
\usage{
get_mutation_rates(object)
}
\arguments{
\item{object}{cevodata object  with fitted cevomod models}
}
\description{
Get mutation rates by Williams
}
