% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_SFS.R
\name{get_non_zero_SFS_range}
\alias{get_non_zero_SFS_range}
\title{Get range of non empty SFS bins}
\usage{
get_non_zero_SFS_range(
  sfs,
  allowed_zero_bins = 1,
  y_treshold = 1,
  y_threshold_pct = 0.01
)
}
\arguments{
\item{sfs}{SFS}

\item{allowed_zero_bins}{number of allowed empty bins in the interval}

\item{y_treshold}{bins with less mutations will be considered empty}

\item{y_threshold_pct}{bins that have less mutations than this param times the
height of the higherst peak will be considered empty}
}
\description{
Get range of non empty SFS bins
}
\keyword{internal}
