% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_patient_mutation_burden}
\alias{get_patient_mutation_burden}
\title{Get tumor mutation burden}
\usage{
get_patient_mutation_burden(cd)
}
\arguments{
\item{cd}{cevodata object}
}
\description{
Get tumor mutation burden
}
