% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{get_residuals}
\alias{get_residuals}
\title{Get model residuals}
\usage{
get_residuals(cd, models_name = cd$active_model)
}
\arguments{
\item{cd}{cevodata object}

\item{models_name}{name of the models}
}
\description{
Get model residuals
}
