% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing_depth.R
\name{get_sample_sequencing_depths}
\alias{get_sample_sequencing_depths}
\title{Get sample sequencing depths}
\usage{
get_sample_sequencing_depths(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{other args}
}
\description{
Get sample sequencing depths
}
