% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_joining.R
\name{identify_non_neutral_tail_mutations}
\alias{identify_non_neutral_tail_mutations}
\alias{identify_non_neutral_tail_mutations.cevodata}
\alias{identify_non_neutral_tail_mutations.singlepatient_cevodata}
\title{Get selected mutations based on 2 sample models}
\usage{
identify_non_neutral_tail_mutations(object, ...)

\method{identify_non_neutral_tail_mutations}{cevodata}(
  object,
  sample1 = NULL,
  sample2 = NULL,
  method = "basic",
  verbose = TRUE,
  ...
)

\method{identify_non_neutral_tail_mutations}{singlepatient_cevodata}(
  object,
  sample1 = NULL,
  sample2 = NULL,
  method = "basic",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{object}

\item{...}{other arguments}

\item{sample1}{sample1}

\item{sample2}{sample2}

\item{method}{method}

\item{verbose}{lgl}
}
\description{
Get selected mutations based on 2 sample models
}
