% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevodata-construction.R
\name{init_cevodata}
\alias{init_cevodata}
\title{Create new cevomod dataset object}
\usage{
init_cevodata(
  name,
  genome = "unknown",
  cancer = "unknown",
  snvs = NULL,
  snvs_name = NULL,
  cnvs = NULL,
  cnvs_name = NULL
)
}
\arguments{
\item{name}{dataset name}

\item{genome}{genome version}

\item{cancer}{cancer type from \code{driver_genes} tbl}

\item{snvs}{tibble with SNVs}

\item{snvs_name}{name for SNVs assay}

\item{cnvs}{tibble with CNVs}

\item{cnvs_name}{name for CNVs assay}
}
\value{
\code{cevodata} object
}
\description{
Create new cevomod dataset object
}
