% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_snvs.R
\name{intervalize_mutation_frequencies}
\alias{intervalize_mutation_frequencies}
\title{Intervalize mutation frequencies}
\usage{
intervalize_mutation_frequencies(
  object,
  which_snvs = default_SNVs(object),
  bins = NULL
)
}
\arguments{
\item{object}{cevodata obj}

\item{which_snvs}{which SNVs to use}

\item{bins}{number of VAF interval bins}
}
\description{
Intervalize mutation frequencies
}
