% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intratumor_heterogeneity.R
\name{ITH}
\alias{ITH}
\alias{plot_ITH}
\alias{plot_ITH.cevodata}
\alias{plot.cevo_ITH_tbl}
\alias{estimate_ITH}
\title{Intratumor heterogeneity}
\usage{
plot_ITH(object, ...)

\method{plot_ITH}{cevodata}(object, ...)

\method{plot}{cevo_ITH_tbl}(x, ...)

estimate_ITH(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{not used}
}
\description{
Intratumor heterogeneity
}
