% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_component.R
\name{layer_exponential_fits}
\alias{layer_exponential_fits}
\title{Plot neutral exponential curve on SFS plot}
\usage{
layer_exponential_fits(
  lm_models,
  start = NA,
  end = 1,
  show.legend = FALSE,
  color = "black",
  size = 1,
  ...
)
}
\arguments{
\item{lm_models}{model fits from fit_neutral_lm}

\item{start}{VAF value to start plotting model fit}

\item{end}{VAF value to end plotting model fit=}

\item{show.legend}{show legend?}

\item{color}{line color}

\item{size}{line size}

\item{...}{other arguments passed to geoms}
}
\description{
Plot neutral exponential curve on SFS plot
}
