% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutral_lm.R
\name{layer_lm_fits}
\alias{layer_lm_fits}
\title{Add M(f) ~ 1/f models layer to M(f) ~ 1/f plot}
\usage{
layer_lm_fits(data, ...)
}
\arguments{
\item{data}{tibble with fits from fit_neutral_partial_models()}

\item{...}{other params passed to geom_segment()}
}
\description{
Add M(f) ~ 1/f models layer to M(f) ~ 1/f plot
}
\examples{
Mf_1f <- tcga_brca_test |>
  dplyr::filter(sample_id \%in\% c("TCGA-AC-A23H-01","TCGA-AN-A046-01")) |>
  dplyr::group_by(sample_id) |>
  calc_Mf_1f()
models <- fit_neutral_lm(Mf_1f, rsq_treshold = 0.99)
plot(Mf_1f, from = 0.05, to = 0.4, scale = FALSE) +
  layer_lm_fits(models)
}
