% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNVs_plots.R
\name{layer_mutations}
\alias{layer_mutations}
\title{Adds mutations to SFS plots}
\usage{
layer_mutations(
  genes = NULL,
  drivers = NULL,
  data = NULL,
  color = "black",
  size = 3,
  shape = "impact",
  filter_fun = guess_filter_fun(shape),
  ...
)
}
\arguments{
\item{genes}{list of genes for which mutations should be shown}

\item{drivers}{cancer type, for which mutations in driver genes should be shown. Needs to
be taken from \code{driver_genes}}

\item{data}{mutations tibble}

\item{color}{color}

\item{size}{size}

\item{shape}{"impact" or "variant_classification"?}

\item{filter_fun}{Function for filtering mutations:
variant_classification_filter() or impact_filter()}

\item{...}{other arguments passed to geom_point()}
}
\value{
ggplot obj
}
\description{
Adds mutations to SFS plots
}
