% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{layer_neutral_tail}
\alias{layer_neutral_tail}
\title{layer_neutral_tail}
\usage{
layer_neutral_tail(object, ...)
}
\arguments{
\item{object}{SNVs tibble object}

\item{...}{other arguments}
}
\description{
layer_neutral_tail
}
