% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_and_themes.R
\name{list_continuous_palettes}
\alias{list_continuous_palettes}
\title{Show available palettes from paletteer}
\usage{
list_continuous_palettes(packages = c("gameofthrones", "harrypotter"))
}
\arguments{
\item{packages}{list of packages}
}
\value{
tbl with palettes
}
\description{
Show available palettes from paletteer
}
