% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_and_themes.R
\name{list_discrete_palettes}
\alias{list_discrete_palettes}
\title{Show available palettes from paletteer}
\usage{
list_discrete_palettes(packages = c("PNWColors"))
}
\arguments{
\item{packages}{list of packages}
}
\value{
tbl with palettes
}
\description{
Show available palettes from paletteer
}
