% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_and_themes.R
\name{list_pnw_palettes}
\alias{list_pnw_palettes}
\title{Show available palettes from PNWColorspackage}
\usage{
list_pnw_palettes()
}
\value{
palette names
}
\description{
Show available palettes from PNWColorspackage
}
