% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevodata-transformations.R
\name{merge.cevodata}
\alias{merge.cevodata}
\title{Merge two cevodata objects}
\usage{
\method{merge}{cevodata}(x, y, name = "Merged datasets", verbose = TRUE, .id = NULL, ...)
}
\arguments{
\item{x, y}{data frames, or objects to be coerced to one.}

\item{name}{Name of the merged object}

\item{verbose}{Show messages?}

\item{.id}{datasets names will be saved to this metadata column, if provided}

\item{...}{arguments to be passed to or from methods.}
}
\description{
Merge two cevodata objects
}
