% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_williams_Mf_1f.R
\name{Mf_1f}
\alias{Mf_1f}
\alias{calc_Mf_1f}
\alias{calc_Mf_1f.cevodata}
\alias{calc_Mf_1f.cevo_snvs}
\alias{plot_Mf_1f}
\alias{plot_Mf_1f.cevodata}
\alias{get_Mf_1f}
\alias{get_Mf_1f.cevodata}
\title{William's M(f) ~ 1/f statistics}
\usage{
calc_Mf_1f(object, ...)

\method{calc_Mf_1f}{cevodata}(object, bins = 100, verbose = TRUE, ...)

\method{calc_Mf_1f}{cevo_snvs}(object, bins = 100, verbose = TRUE, ...)

plot_Mf_1f(object, ...)

\method{plot_Mf_1f}{cevodata}(
  object,
  bins = NULL,
  from = 0.1,
  to = 0.25,
  scale = TRUE,
  geom = "point",
  ...
)

get_Mf_1f(object, ...)

\method{get_Mf_1f}{cevodata}(object, model_name = "Mf_1f", verbose = TRUE, ...)
}
\arguments{
\item{object}{SNVs tibble object}

\item{...}{other arguments}

\item{bins}{resolution of the cumulative tails calculation}

\item{verbose}{verbose?}

\item{from}{min VAF to plot}

\item{to}{max VAF to plot}

\item{scale}{scale data?}

\item{geom}{ggplot geom to use, eg. geom_line()}

\item{model_name}{name of slot with Mf_1f statistics}
}
\value{
ggplot obj
}
\description{
Creates  cevodata$models$Mf_1f tbl with the groupping variables and:
\itemize{
\item VAF
\item n - number of mutations in the VAF interval
\item \code{M(f)} and \code{1/f} columns to plot William's statistics
}
}
\section{Functions}{
\itemize{
\item \code{calc_Mf_1f()}: Calculate Williams M(f) ~ 1/f

\item \code{calc_Mf_1f(cevodata)}: Calculate Williams M(f) ~ 1/f

\item \code{calc_Mf_1f(cevo_snvs)}: Calculate Williams M(f) ~ 1/f

\item \code{plot_Mf_1f(cevodata)}: Plot M(f) ~ 1/f

\item \code{get_Mf_1f()}: Get Mf_1f

\item \code{get_Mf_1f(cevodata)}: Get Mf_1f

}}
\examples{
data("tcga_brca_test")
tcga_brca_test |>
  calc_Mf_1f()

tcga_brca_test |>
  plot_Mf_1f()
}
