% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNVs_plots.R, R/generics.R
\name{plot_mutations.cevodata}
\alias{plot_mutations.cevodata}
\alias{plot_mutations.tbl_df}
\alias{layer_mutations}
\alias{mutation_plots}
\alias{plot_mutations}
\title{Show mutations in particular genes}
\usage{
\method{plot_mutations}{cevodata}(
  object,
  genes = NULL,
  drivers = NULL,
  mark_genes = NULL,
  y = "genes",
  shape = "impact",
  filter_fun = guess_filter_fun(shape),
  ...
)

\method{plot_mutations}{tbl_df}(
  object,
  genes = NULL,
  drivers = NULL,
  mark_genes = NULL,
  y = "genes",
  shape = "impact",
  filter_fun = guess_filter_fun(shape),
  ...
)

layer_mutations(
  genes = NULL,
  drivers = NULL,
  data = NULL,
  color = "black",
  size = 3,
  shape = "impact",
  filter_fun = guess_filter_fun(shape),
  ...
)

plot_mutations(object, ...)
}
\arguments{
\item{object}{cevodata}

\item{genes}{list of genes for which mutations should be shown}

\item{drivers}{cancer type, for which mutations in driver genes should be shown. Needs to
be taken from \code{driver_genes}}

\item{mark_genes}{list of genes to mark}

\item{y}{which to show on y axis: "genes" or "samples"}

\item{shape}{"impact" or "variant_classification"?}

\item{filter_fun}{Function for filtering mutations:
variant_classification_filter() or impact_filter()}

\item{...}{other arguments passed to geom_point()}

\item{color}{color}

\item{size}{size}
}
\value{
ggplot obj
}
\description{
Show mutations in particular genes
}
\section{Functions}{
\itemize{
\item \code{layer_mutations()}: Adds mutations to SFS plots

}}
