% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutral_lm.R
\name{neutral_lm}
\alias{neutral_lm}
\alias{fit_neutral_lm}
\alias{fit_neutral_lm.cevodata}
\alias{layer_lm_fits}
\title{Fitting neutral models}
\usage{
fit_neutral_lm(object, ...)

\method{fit_neutral_lm}{cevodata}(object, rsq_treshold = 0.98, ...)

layer_lm_fits(cd, ...)
}
\arguments{
\item{object}{SNVs tibble object}

\item{...}{other params passed to geom_segment()}

\item{rsq_treshold}{R-squared tresholds to keep model as neutral}

\item{cd}{cevodata}
}
\description{
Creates  cevodata$models$neutral_lm
}
\section{Functions}{
\itemize{
\item \code{fit_neutral_lm(cevodata)}: Fit Williams neutral models to the data

\item \code{layer_lm_fits()}: Add M(f) ~ 1/f models layer to M(f) ~ 1/f plot

}}
\examples{
data("tcga_brca_test")
snvs <- SNVs(tcga_brca_test) |>
  dplyr::filter(sample_id \%in\% c("TCGA-AC-A23H-01","TCGA-AN-A046-01"))

cd <- init_cevodata("Test") |>
  add_SNV_data(snvs) |>
  calc_Mf_1f() |>
  calc_SFS() |>
  fit_neutral_lm(rsq_treshold = 0.99)

plot(cd$models$Mf_1f, from = 0.05, to = 0.4, scale = FALSE) +
  layer_lm_fits(cd)
}
