% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_cumulative_tails.R
\name{plot.cevo_cumulative_tails_tbl}
\alias{plot.cevo_cumulative_tails_tbl}
\title{Plot the cumulative tails}
\usage{
\method{plot}{cevo_cumulative_tails_tbl}(x, scale_y = TRUE, scales = "loglog", ...)
}
\arguments{
\item{x}{tibble with calc_cumulative_tails() results}

\item{scale_y}{logical}

\item{scales}{loglog/semilog}

\item{...}{required by Generic}
}
\value{
ggplot obj
}
\description{
Plot the cumulative tails
}
\examples{
data("tcga_brca_test")
SNVs(tcga_brca_test) |>
  dplyr::group_by(sample_id) |>
  calc_cumulative_tails() |>
  plot()
}
