% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/williams_Mf_1f.R
\name{plot.cevo_Mf_1f_tbl}
\alias{plot.cevo_Mf_1f_tbl}
\alias{plot_Mf_1f}
\title{Plot M(f) ~ 1/f}
\usage{
\method{plot}{cevo_Mf_1f_tbl}(
  x,
  from = 0.1,
  to = 0.25,
  scale = TRUE,
  geom = "point",
  mapping = NULL,
  ...
)

plot_Mf_1f(
  dt,
  digits = 2,
  from = 0.1,
  to = 0.25,
  scale = TRUE,
  geom = "point",
  ...
)
}
\arguments{
\item{x}{tibble with calc_Mf_1f() results}

\item{from}{min VAF to plot}

\item{to}{max VAF to plot}

\item{scale}{scale data?}

\item{geom}{ggplot geom to use, eg. geom_line()}

\item{mapping}{mapping}

\item{...}{futher passed to geom_()}

\item{dt}{tibble with VAF column. Might be groupped and contain the data for multiple samples}

\item{digits}{resolution of the cumulative tails calculation}
}
\value{
ggplot obj
}
\description{
Plot M(f) ~ 1/f
}
\section{Functions}{
\itemize{
\item \code{plot_Mf_1f()}: Plot M(f) ~ 1/f

}}
\examples{
data("tcga_brca_test")
tcga_brca_test |>
  dplyr::group_by(sample_id) |>
  calc_Mf_1f() |>
  plot()

tcga_brca_test |>
  dplyr::group_by(sample_id) |>
  plot_Mf_1f()
}
