% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_tails.R
\name{plot_cumulative_tails}
\alias{plot_cumulative_tails}
\title{Plot the cumulative tails}
\usage{
plot_cumulative_tails(dt, scale_y = TRUE, ...)
}
\arguments{
\item{dt}{snvs_ tibble with VAF and sample_id columns}

\item{scale_y}{scale y vaules to 1?}

\item{...}{passed to stat_cumulative_tail}
}
\value{
ggplot obj
}
\description{
Shortcut to plot cum tails from snvs_ dataframe
}
\examples{
data("tcga_brca_test")
tcga_brca_test |>
  plot_cumulative_tails()
}
