% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutral_lm.R
\name{plot_lm_a_coefficients}
\alias{plot_lm_a_coefficients}
\title{Plot 'a' coefficients for all fitted neutral models}
\usage{
plot_lm_a_coefficients(object, ...)
}
\arguments{
\item{object}{cevodata object}

\item{...}{other parameters passed to geom}
}
\description{
Plot 'a' coefficients for all fitted neutral models
}
