% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNVs_plots.R
\name{plot_mutations}
\alias{plot_mutations}
\title{Show mutations in particular genes}
\usage{
plot_mutations(
  data,
  genes = NULL,
  drivers = NULL,
  mark_genes = NULL,
  y = "genes",
  shape = "impact",
  filter_fun = guess_filter_fun(shape)
)
}
\arguments{
\item{data}{mutations tibble}

\item{genes}{list of genes for which mutations should be shown}

\item{drivers}{cancer type, for which mutations in driver genes should be shown. Needs to
be taken from \code{driver_genes}}

\item{mark_genes}{list of genes to mark}

\item{y}{which to show on y axis: "genes" or "samples"}

\item{shape}{"impact" or "variant_classification"?}

\item{filter_fun}{Function for filtering mutations:
variant_classification_filter() or impact_filter()}
}
\value{
ggplot obj
}
\description{
Show mutations in particular genes
}
