% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_joining.R
\name{plot_non_neutral_mutations_2D}
\alias{plot_non_neutral_mutations_2D}
\alias{plot_non_neutral_mutations_2D.cevodata}
\title{Plot heatmap of non neutral mutations}
\usage{
plot_non_neutral_mutations_2D(object, ...)

\method{plot_non_neutral_mutations_2D}{cevodata}(object, colors = c("black", "white", "red"), ...)
}
\arguments{
\item{object}{object}

\item{...}{other arguments}

\item{colors}{vector of three colors to use}
}
\description{
Plot heatmap of non neutral mutations
}
