% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_data_overview.R
\name{plot_private_shared_mutations}
\alias{plot_private_shared_mutations}
\alias{plot_private_shared_mutations.cevodata}
\title{Plot proportions of private and shared mutations}
\usage{
plot_private_shared_mutations(object, ...)

\method{plot_private_shared_mutations}{cevodata}(object, mapping = NULL, geom = geom_bar, ...)
}
\arguments{
\item{object}{cevodata}

\item{...}{other params}

\item{mapping}{customize mapping}

\item{geom}{change geom}
}
\description{
Plot proportions of private and shared mutations
}
\section{Methods (by class)}{
\itemize{
\item \code{plot_private_shared_mutations(cevodata)}: Plot private and shared mutation fractions

}}
