% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_data_overview.R
\name{plot_sequencing_depth}
\alias{plot_sequencing_depth}
\alias{plot_sequencing_depth.cevodata}
\alias{plot_sequencing_depth.cevo_snvs}
\title{Plot sequencing depth of cevodata}
\usage{
plot_sequencing_depth(object, ...)

\method{plot_sequencing_depth}{cevodata}(
  object,
  snvs_name = default_SNVs(object),
  geom = geom_boxplot,
  ...
)

\method{plot_sequencing_depth}{cevo_snvs}(object, geom = geom_boxplot, ...)
}
\arguments{
\item{object}{cevodata or cevo_snvs}

\item{...}{params passed to geom}

\item{snvs_name}{name of the SNVs table}

\item{geom}{change geom}
}
\description{
Plot sequencing depth of cevodata
}
\section{Methods (by class)}{
\itemize{
\item \code{plot_sequencing_depth(cevodata)}: Plot sequencing data of mutations in cevodata

\item \code{plot_sequencing_depth(cevo_snvs)}: Plot sequencing data of mutations in cevodata

}}
