% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_snvs.R
\name{prepare_SNVs}
\alias{prepare_SNVs}
\title{Prepare SNVs for analyses}
\usage{
prepare_SNVs(object, which = default_SNVs(object), bins = NULL, verbose = TRUE)
}
\arguments{
\item{object}{cevodata obj}

\item{which}{which SNVs to use}

\item{bins}{number of VAF interval bins}

\item{verbose}{verbose?}
}
\description{
Prepare SNVs for analyses
}
