% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_and_themes.R
\name{print.cevo_palettes}
\alias{print.cevo_palettes}
\title{Show implemented palettes}
\usage{
\method{print}{cevo_palettes}(x, ...)
}
\arguments{
\item{x}{Palettes}

\item{...}{other args}
}
\description{
Show implemented palettes
}
