% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_and_themes.R
\name{print_palettes}
\alias{print_palettes}
\title{Show palettes}
\usage{
print_palettes(packages = c("PNWColors", "nord"), n = 10)
}
\arguments{
\item{packages}{list of packages}

\item{n}{number of colors}
}
\description{
Show palettes
}
