% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevodata-readthis_integration.R
\name{readthis-integration}
\alias{readthis-integration}
\alias{add_data}
\alias{add_to_cevodata}
\title{readthis integration}
\usage{
add_data(cd, data, ...)

add_to_cevodata(data, cd, name, verbose, ...)
}
\arguments{
\item{cd}{\if{html}{\out{<cevodata>}} object}

\item{data}{Object read with readthis functions}

\item{...}{Other arguments}

\item{name}{Name for the data}

\item{verbose}{Verbose?}
}
\description{
\href{https://github.com/pawelqs/readthis}{readthis} package may be used to easily
read the data from some popular mutation callers into R environment. readthis
functions can be supplied not only with the single file paths, but also with
lists of files or even paths to the directories with files to be loaded (and
cevodata object is to store the data from many samples!)

readthis functions return tibbles or list of tibbles. These tibbles/
objects usually are instances of \emph{cevo_<software_name>} S3 classes. cevomod
implements methods that allow to add these types of data to the cevodata
objects conveniently.
}
\section{Functions}{
\itemize{
\item \code{add_data()}: add_data() function takes cevodata as the
first argument, so it is a preferred method for adding data in R pipelines.

\item \code{add_to_cevodata()}: add_to_cevodata() is a generic with a set
of methods for different classes of \code{data}. These methods are called by
add_data() function.

}}
\examples{
# library(cevomod)

ascat_dir <- system.file("extdata", "ASCAT", package = "readthis")
ascat <- readthis::read_ascat_files(ascat_dir)
cd <- init_cevodata("Test dataset") |>
  add_data(ascat)

}
