% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_and_themes.R
\name{rotate_x_labels}
\alias{rotate_x_labels}
\title{Rotate x axix labels}
\usage{
rotate_x_labels(angle = 90, vjust = 0.5)
}
\arguments{
\item{angle}{angle}

\item{vjust}{vjust}
}
\description{
Rotate x axix labels
}
