% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_and_themes.R
\name{scale_fill_cevomod}
\alias{scale_fill_cevomod}
\alias{scale_color_cevomod}
\title{Different color pallettes}
\usage{
scale_fill_cevomod(palette = "starfleet", ...)

scale_color_cevomod(palette = "starfleet", ...)
}
\arguments{
\item{palette}{palette name}

\item{...}{other arguments passed to scale_*_manual()}
}
\description{
Different color pallettes
}
