% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_and_themes.R
\name{scale_fill_pnw}
\alias{scale_fill_pnw}
\title{Use PNWColors palette}
\usage{
scale_fill_pnw(palette = "Sunset", direction = 1, dynamic = FALSE, ...)
}
\arguments{
\item{palette}{palette name}

\item{direction}{1/-1}

\item{dynamic}{TRUE/FALSE}

\item{...}{additional arguments to pass to discrete_scale}
}
\description{
Use PNWColors palette
}
