% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevodata-construction.R
\name{set_cancer_type}
\alias{set_cancer_type}
\alias{set_cancer_type.cevodata}
\title{Set cancer type for the object}
\usage{
set_cancer_type(object, ...)

\method{set_cancer_type}{cevodata}(object, cancer_type, ...)
}
\arguments{
\item{cancer_type}{cancer type}
}
\description{
Set cancer type for the object
}
\section{Methods (by class)}{
\itemize{
\item \code{set_cancer_type(cevodata)}: Set cancer type for cevodata object

}}
