% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevomod-persistent.R
\name{set_containers_dir}
\alias{set_containers_dir}
\alias{get_containers_dir}
\title{Get/Set the containers directory}
\usage{
set_containers_dir(dir)

get_containers_dir()
}
\arguments{
\item{dir}{Path for containers}
}
\description{
Get/Set the containers directory
}
