% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shuffle}
\alias{shuffle}
\alias{shuffle.tbl_df}
\title{Shuffle order of elements in object}
\usage{
shuffle(object, ...)

\method{shuffle}{tbl_df}(object, ...)
}
\arguments{
\item{object}{object to shuffle}

\item{...}{other arguments}
}
\description{
Shuffle order of elements in object
}
\section{Methods (by class)}{
\itemize{
\item \code{shuffle(tbl_df)}: Shuffle order of rows in tibble

}}
\examples{
tibble::tibble(i = 1:10) |>
  shuffle()
}
