% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevodata-getters.R
\name{SNVs_CNVs}
\alias{SNVs_CNVs}
\title{Get SNVs with merged CNVs}
\usage{
SNVs_CNVs(object)
}
\arguments{
\item{object}{cevodata object with SNVs and CNVs}
}
\description{
Get SNVs with merged CNVs
}
