% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevodata-transformations.R
\name{split_by.cevodata}
\alias{split_by.cevodata}
\title{Split cevodata object}
\usage{
\method{split_by}{cevodata}(object, x, ...)
}
\arguments{
\item{object, }{cevodata object}

\item{x}{name of column in metadata}
}
\description{
Split cevodata object
}
