% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevodata-transformations.R
\name{split_by}
\alias{split_by}
\alias{split_by.cevodata}
\title{Split object}
\usage{
split_by(object, ...)

\method{split_by}{cevodata}(object, x, ...)
}
\arguments{
\item{object}{cevodata object}

\item{...}{other arguments}

\item{x}{name of column in metadata}
}
\description{
Split object
}
\section{Methods (by class)}{
\itemize{
\item \code{split_by(cevodata)}: Split cevodata object

}}
