% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFS.R
\name{stat_SFS}
\alias{stat_SFS}
\title{SFS stat}
\usage{
stat_SFS(
  mapping = NULL,
  data = NULL,
  binwidth = 0.01,
  geom = "line",
  position = "identity",
  ...
)
}
\arguments{
\item{mapping}{aes()}

\item{data}{data}

\item{binwidth}{binwidth}

\item{geom}{geom}

\item{position}{position}

\item{...}{args passed to stat_bin}
}
\description{
Modification of stat_bin with custom parameters
}
