% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_tung_durrett_model.R
\name{tung_durrett}
\alias{tung_durrett}
\alias{fit_tung_durrett_models}
\alias{fit_tung_durrett_models.cevodata}
\title{Fitting Tung Durrett models}
\usage{
fit_tung_durrett_models(object, ...)

\method{fit_tung_durrett_models}{cevodata}(object, name = "tung_durrett", verbose = TRUE, ...)
}
\arguments{
\item{object}{cevodata}

\item{...}{other arguments}

\item{name}{name in the models' slot}

\item{verbose}{verbose?}
}
\description{
Fitting Tung Durrett models
}
\examples{
data("tcga_brca_test")
cd <- tcga_brca_test |>
  dplyr::filter(sample_id \%in\% c("TCGA-AC-A23H-01","TCGA-AN-A046-01")) |>
  fit_tung_durrett_models()
}
