% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevo_snvs.R
\name{unite_mutation_id}
\alias{unite_mutation_id}
\title{Unite many columns to create mutation_id column}
\usage{
unite_mutation_id(snvs, sep = "-", remove = TRUE)
}
\arguments{
\item{snvs}{SNVs}

\item{sep}{Separator}

\item{remove}{Remove united columns?}
}
\description{
Unite many columns to create mutation_id column
}
