% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevodata-transformations.R
\name{update.cevodata}
\alias{update.cevodata}
\title{Update cevodata object with values from another object}
\usage{
\method{update}{cevodata}(object, object2, ...)
}
\arguments{
\item{object}{object to update}

\item{object2}{object to use}

\item{...}{other args, unused now}
}
\description{
Update cevodata object with values from another object
}
