% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevodata-construction.R
\name{use_purity}
\alias{use_purity}
\title{Choose purity measure}
\usage{
use_purity(cd, name, verbose = get_cevomod_verbosity())
}
\arguments{
\item{cd}{\if{html}{\out{<cevodata>}} object}

\item{name}{Name of the metadata column with chosen purity values}

\item{verbose}{Verbose?}
}
\description{
\if{html}{\out{<cevodata>}} metadata can contain purity measures in columns other than 'purity'.
T his function can be used to set 'purity' values using values from requested
column
}
