% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_plots.R
\name{variant_classification_filter}
\alias{variant_classification_filter}
\alias{impact_filter}
\title{Filter mutations by variant_classification field}
\usage{
variant_classification_filter(keep = c("nonsilent", "null"))

impact_filter(keep = c("HIGH", "MODERATE"))
}
\arguments{
\item{keep}{values to keep}
}
\value{
filtering function
}
\description{
Filter mutations by variant_classification field
}
\section{Functions}{
\itemize{
\item \code{impact_filter()}: Filter mutations by impact field

}}
