% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_williams_neutral_model.R
\name{williams_neutral_model}
\alias{williams_neutral_model}
\alias{fit_williams_neutral_models}
\alias{fit_williams_neutral_models.cevodata}
\alias{layer_lm_fits}
\title{Fitting neutral models}
\usage{
fit_williams_neutral_models(object, ...)

\method{fit_williams_neutral_models}{cevodata}(
  object,
  rsq_treshold = 0.98,
  name = "williams_neutral",
  verbose = TRUE,
  ...
)

layer_lm_fits(cd, model_name = "williams_neutral", ...)
}
\arguments{
\item{object}{SNVs tibble object}

\item{...}{other params passed to geom_segment()}

\item{rsq_treshold}{R-squared tresholds to keep model as neutral}

\item{name}{name in the models' slot}

\item{verbose}{verbose?}

\item{cd}{cevodata}

\item{model_name}{modelname}
}
\description{
Creates  cevodata$models$neutral_model
}
\section{Functions}{
\itemize{
\item \code{layer_lm_fits()}: Add M(f) ~ 1/f models layer to M(f) ~ 1/f plot

}}
\examples{
data("tcga_brca_test")
snvs <- SNVs(tcga_brca_test) |>
  dplyr::filter(sample_id \%in\% c("TCGA-AC-A23H-01","TCGA-AN-A046-01"))

cd <- init_cevodata("Test") |>
  add_SNV_data(snvs) |>
  calc_Mf_1f() |>
  calc_SFS() |>
  fit_williams_neutral_models(rsq_treshold = 0.99)

plot(cd$models$Mf_1f, from = 0.05, to = 0.4, scale = FALSE) +
  layer_lm_fits(cd)
}
