% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevodata-class.R, R/generics.R
\name{default_SNVs.cevodata}
\alias{default_SNVs.cevodata}
\alias{default_SNVs<-.cevodata}
\alias{default_CNVs.cevodata}
\alias{default_CNVs<-.cevodata}
\alias{active_assays}
\alias{default_SNVs}
\alias{default_SNVs<-}
\alias{default_CNVs}
\alias{default_CNVs<-}
\title{Get/Set active assays of the cevodata object}
\usage{
\method{default_SNVs}{cevodata}(object, ...)

\method{default_SNVs}{cevodata}(object, ...) <- value

\method{default_CNVs}{cevodata}(object, ...)

\method{default_CNVs}{cevodata}(object, ...) <- value

default_SNVs(object, ...)

default_SNVs(object, ...) <- value

default_CNVs(object, ...)

default_CNVs(object, ...) <- value
}
\arguments{
\item{object}{object}

\item{...}{other arguments}

\item{value}{name of new default assay}
}
\description{
Get/Set active assays of the cevodata object
}
\section{Functions}{
\itemize{
\item \code{default_SNVs(cevodata)}: Get default SNVs assay of cevodata

\item \code{default_SNVs(cevodata) <- value}: Set default SNVs assay of cevodata

\item \code{default_CNVs(cevodata)}: Get default CNVs assay of cevodata

\item \code{default_CNVs(cevodata) <- value}: Set default CNVs assay of cevodata

}}
