% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containers.R
\name{build_CliP_container}
\alias{build_CliP_container}
\title{Build Apptainer container for CliP}
\usage{
build_CliP_container(out_dir = NULL, force = FALSE)
}
\arguments{
\item{out_dir}{path to save the CliP.sif file}
}
\description{
Build Apptainer container for CliP
}
