% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevodata-construction.R
\name{cevo_metadata}
\alias{cevo_metadata}
\alias{add_patient_data}
\alias{add_patient_data.cevodata}
\alias{add_sample_data}
\alias{add_sample_data.cevodata}
\title{Add metadata to the cevodata object}
\usage{
add_patient_data(object, ...)

\method{add_patient_data}{cevodata}(object, data, ...)

add_sample_data(object, ...)

\method{add_sample_data}{cevodata}(object, data, ...)
}
\arguments{
\item{object}{object}

\item{...}{other arguments}

\item{data}{name of new default assay}
}
\description{
Add metadata to the cevodata object
}
\section{Functions}{
\itemize{
\item \code{add_patient_data(cevodata)}: Add patient data to cevodata object

\item \code{add_sample_data(cevodata)}: Add samples' data to cevodata object

}}
