% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_muts_by_component.R
\name{count_mutations_by_component}
\alias{count_mutations_by_component}
\alias{count_powerlaw_tail_mutations}
\alias{count_missing_powerlaw_tail_mutations}
\title{Get numbers of neutral, clonal, subclonal ect variants}
\usage{
count_mutations_by_component(
  object,
  models_name = active_models(object),
  include_missing = FALSE,
  min_f = 0.01
)

count_powerlaw_tail_mutations(object, models_name = active_models(object))

count_missing_powerlaw_tail_mutations(
  object,
  models_name = active_models(object),
  min_f = 0.01
)
}
\arguments{
\item{object}{cevodata object}

\item{models_name}{models name}

\item{include_missing}{include the component of the missing mutations}

\item{min_f}{f threshold - power law tail goes to the Infinity, so we need
to set a reasonable threshold}
}
\description{
Get numbers of neutral, clonal, subclonal ect variants
}
\section{Functions}{
\itemize{
\item \code{count_powerlaw_tail_mutations()}: Count neutral tail mutations

\item \code{count_missing_powerlaw_tail_mutations()}: Count filtered mutations with f higher than threshold

}}
