% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_tails.R, R/generics.R
\name{calc_cumulative_tails.cevodata}
\alias{calc_cumulative_tails.cevodata}
\alias{calc_cumulative_tails.tbl_df}
\alias{plot_cumulative_tails.cevodata}
\alias{cumulative_tails}
\alias{calc_cumulative_tails}
\alias{plot_cumulative_tails}
\title{Cumulative tails}
\usage{
\method{calc_cumulative_tails}{cevodata}(object, digits = 2, ...)

\method{calc_cumulative_tails}{tbl_df}(object, digits = 2, ...)

\method{plot_cumulative_tails}{cevodata}(object, scale_y = TRUE, ...)

calc_cumulative_tails(object, ...)

plot_cumulative_tails(object, ...)
}
\arguments{
\item{object}{SNVs tibble object}

\item{digits}{resolution of the cumulative tails calculation}

\item{...}{other arguments}

\item{scale_y}{scale y vaules to 1?}
}
\value{
ggplot obj
}
\description{
Creates  cevodata$models$cumulative_tails tbl with the groupping variables and:
\itemize{
\item n column with the number of mutations in the VAF interval
\item x and y columns describing the cumulative tails
\item y_scaled with y values scaled to the range 0-1
}
}
\section{Functions}{
\itemize{
\item \code{calc_cumulative_tails(cevodata)}: Calculate the cumulative tails

\item \code{calc_cumulative_tails(tbl_df)}: Calculate the cumulative tails

\item \code{plot_cumulative_tails(cevodata)}: Shortcut to plot cum tails from SNVs dataframe

}}
\examples{
data("tcga_brca_test")
tcga_brca_test |>
  calc_cumulative_tails()

tcga_brca_test |>
  plot_cumulative_tails()
}
