% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_cumulative_tails.R
\name{cumulative_tails}
\alias{cumulative_tails}
\alias{calc_cumulative_tails}
\alias{calc_cumulative_tails.cevodata}
\alias{calc_cumulative_tails.cevo_snvs}
\alias{plot_cumulative_tails}
\alias{plot_cumulative_tails.cevodata}
\title{Cumulative tails}
\usage{
calc_cumulative_tails(object, ...)

\method{calc_cumulative_tails}{cevodata}(object, bins = 100, ...)

\method{calc_cumulative_tails}{cevo_snvs}(object, bins = 100, ...)

plot_cumulative_tails(object, ...)

\method{plot_cumulative_tails}{cevodata}(object, scale_y = TRUE, scales = "loglog", ...)
}
\arguments{
\item{object}{SNVs tibble object}

\item{...}{passed to stat_cumulative_tail}

\item{scale_y}{scale y vaules to 1?}

\item{digits}{resolution of the cumulative tails calculation}
}
\value{
ggplot obj
}
\description{
Creates  cevodata$models$cumulative_tails tbl with the groupping variables and:
\itemize{
\item n column with the number of mutations in the VAF interval
\item x and y columns describing the cumulative tails
\item y_scaled with y values scaled to the range 0-1
}
}
\section{Functions}{
\itemize{
\item \code{calc_cumulative_tails(cevodata)}: Calculate the cumulative tails

\item \code{calc_cumulative_tails(cevo_snvs)}: Calculate the cumulative tails

\item \code{plot_cumulative_tails(cevodata)}: Shortcut to plot cum tails from SNVs dataframe

}}
\examples{
data("tcga_brca_test")
tcga_brca_test |>
  calc_cumulative_tails()

tcga_brca_test |>
  plot_cumulative_tails()
}
