% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_cumulative_tails.R
\name{cumulative_tails}
\alias{cumulative_tails}
\alias{calc_cumulative_tails}
\alias{calc_cumulative_tails.cevodata}
\alias{calc_cumulative_tails.cevo_snvs}
\alias{plot_cumulative_tails}
\alias{plot_cumulative_tails.cevodata}
\title{Cumulative tails}
\usage{
calc_cumulative_tails(object, ...)

\method{calc_cumulative_tails}{cevodata}(
  object,
  which_snvs = default_SNVs(object),
  column = get_frequency_measure_name(object, which_snvs),
  bins = 100,
  verbose = get_cevomod_verbosity(),
  ...
)

\method{calc_cumulative_tails}{cevo_snvs}(
  object,
  column = get_frequency_measure_name(object),
  bins = 100,
  verbose = get_cevomod_verbosity(),
  ...
)

plot_cumulative_tails(object, ...)

\method{plot_cumulative_tails}{cevodata}(object, scale_y = TRUE, scales = "loglog", ...)
}
\arguments{
\item{object}{SNVs tibble object}

\item{...}{passed to geom_line()}

\item{which_snvs}{Which SNVs to use?}

\item{column}{VAF or CCF/2}

\item{bins}{Resolution of the cumulative tails calculation}

\item{verbose}{Verbose?}

\item{scale_y}{scale y vaules to 1?}

\item{scales}{loglog/semilog}
}
\value{
ggplot obj
}
\description{
cumulative_tails columns:
\itemize{
\item f
\item n column with the number of mutations in the f interval
\item y cumulative tail value
\item y_scaled with y values scaled to the range 0-1
}
}
\section{Functions}{
\itemize{
\item \code{calc_cumulative_tails(cevodata)}: Calculate the cumulative tails and saves to
cevodata$models$cumulative_tails tibble

\item \code{calc_cumulative_tails(cevo_snvs)}: Calculate the cumulative tails

\item \code{plot_cumulative_tails(cevodata)}: Shortcut to plot cum tails from SNVs dataframe

}}
\examples{
data("test_data")
test_data |>
  calc_cumulative_tails()

test_data |>
  plot_cumulative_tails()
}
