% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{cut_f_intervals}
\alias{cut_f_intervals}
\alias{cut_f_intervals.cevo_snvs}
\title{Cut f or VAF into intervals}
\usage{
cut_f_intervals(object, ...)

\method{cut_f_intervals}{cevo_snvs}(object, bins = NULL, column = "VAF", ...)
}
\arguments{
\item{object}{object}

\item{...}{other params}

\item{bins}{number of bins}
}
\description{
Cut f or VAF into intervals
}
