% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlaw_neutral_tail.R
\name{estimate_sampling_rate}
\alias{estimate_sampling_rate}
\title{Estimate sampling rate}
\usage{
estimate_sampling_rate(sfs, lm_models)
}
\arguments{
\item{sfs}{SFS}

\item{lm_models}{output from fit_neutral_lm()}
}
\value{
tibble
}
\description{
Uses experimental SFS and power-law model to estimate the sampling rate
}
