% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolutionary_parameters.R
\name{evo_params}
\alias{evo_params}
\alias{get_mutation_rates}
\alias{get_mutation_rates.cevodata}
\alias{get_mutation_rates.tbl_df}
\alias{get_selection_coefficients}
\alias{get_selection_coefficients.cevodata}
\alias{get_selection_coefficients.tbl_df}
\title{Get evolutionary parameters from the model}
\usage{
get_mutation_rates(object, ...)

\method{get_mutation_rates}{cevodata}(object, models_name = "powerlaw_fixed", ...)

\method{get_mutation_rates}{tbl_df}(object, ...)

get_selection_coefficients(object, ...)

\method{get_selection_coefficients}{cevodata}(
  object,
  models_name = "williams_neutral_subclones",
  Nmax = 10^10,
  ...
)

\method{get_selection_coefficients}{tbl_df}(object, Nmax = 10^10, ...)
}
\arguments{
\item{object}{cevodata object or models tibble}

\item{...}{other arguments}

\item{models_name}{models_name}

\item{Nmax}{Time when tumour is sampled (in tumour doublings)}
}
\description{
Code of those functions is only re-formatted code from MOBSTER R package
by Caravagna, Williams et al. https://github.com/caravagnalab/mobster
}
\section{Functions}{
\itemize{
\item \code{get_mutation_rates()}: Get mutation rates by Williams
Use ...

\item \code{get_mutation_rates(cevodata)}: Get mutation rates by Williams

\item \code{get_mutation_rates(tbl_df)}: Get mutation rates by Williams

\item \code{get_selection_coefficients()}: Get subclonal selection coefficients Williams

Use properties of subclone fit to calculate selection intensity, selection
is defined as the relative growth rates of host tumour cell
populations (\eqn{\lambda h}) vs subclone (\eqn{\lambda s}):
\deqn{1+s=\lambda h / \lambda s}

\item \code{get_selection_coefficients(cevodata)}: Get subclonal selection coefficients Williams

\item \code{get_selection_coefficients(tbl_df)}: Get subclonal selection coefficients Williams

}}
