% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_evolutionary_parameters.R
\name{evo_params}
\alias{evo_params}
\alias{get_evolutionary_parameters}
\title{Get evolutionary parameters from the model}
\usage{
get_evolutionary_parameters(object, ...)
}
\arguments{
\item{object}{cevodata object or models tibble}

\item{...}{other arguments}

\item{Nmax}{Time when tumour is sampled (in tumour doublings)}

\item{models_name}{models_name}
}
\description{
Get evolutionary parameters from the model
}
\details{
Code to get the evolutionary parameters from the powerlaw model with subclones
is only re-formatted code from MOBSTER R package by Caravagna, Williams et al.
https://github.com/caravagnalab/mobster

Uses properties of subclone fit to calculate selection intensity, selection
is defined as the relative growth rates of host tumour cell
populations (\eqn{\lambda h}) vs subclone (\eqn{\lambda s}):
\deqn{1+s=\lambda h / \lambda s}
}
