% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_subclones_binomial_BMix.R,
%   R/fit_subclones_binomial_mclust.R
\name{fit_subclones_bmix}
\alias{fit_subclones_bmix}
\alias{fit_subclones}
\title{Fit subclonal distributions to neutral model residuals}
\usage{
fit_subclones_bmix(
  object,
  N = 1:3,
  snvs_name = default_SNVs(object),
  powerlaw_model_name = active_models(object),
  upper_VAF_limit = 0.75,
  verbose = TRUE
)

fit_subclones(
  object,
  N = 1:3,
  powerlaw_model_name = active_models(object),
  upper_VAF_limit = 0.75,
  verbose = TRUE
)
}
\arguments{
\item{object}{object}

\item{N}{vector of numbers of clones to test}

\item{snvs_name}{which snvs to to use?}

\item{powerlaw_model_name}{residual of which powerlaw model to use?
powerlaw_fixed/powerlaw_optim}

\item{upper_VAF_limit}{ignore variants with f higher than}

\item{verbose}{verbose?}
}
\description{
Fit subclonal distributions to neutral model residuals
}
\section{Functions}{
\itemize{
\item \code{fit_subclones_bmix()}: Fit subclonal distributions to neutral model residuals using BMix

}}
