% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclones_binomial.R
\name{fit_subclones}
\alias{fit_subclones}
\title{Fit subclonal distributions to neutral model residuals}
\usage{
fit_subclones(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{other arguments}
}
\description{
Fit subclonal distributions to neutral model residuals
}
