% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_subclones_binomial.R
\name{fit_subclones}
\alias{fit_subclones}
\alias{fit_subclones.cevodata}
\title{Fit subclonal distributions to neutral model residuals}
\usage{
fit_subclones(object, ...)

\method{fit_subclones}{cevodata}(
  object,
  N = 1:3,
  powerlaw_model_name = active_models(object),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{object}

\item{...}{other arguments}

\item{N}{vector of numbers of clones to test}

\item{verbose}{verbose?}

\item{powerlaw_model}{residual of which powerlaw model to use?
williams_neutral/tung_durrett}
}
\description{
Fit subclonal distributions to neutral model residuals
}
