% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_subclones.R,
%   R/fit_subclones_binomial_BMix.R, R/fit_subclones_binomial_mclust.R
\name{fit_subclones}
\alias{fit_subclones}
\alias{fit_subclones_bmix}
\alias{fit_subclones_mclust}
\title{Fit clonal and subclonal components of the model to the residuals of the
power-law model}
\usage{
fit_subclones(
  object,
  N = 1:3,
  powerlaw_model_name = active_models(object),
  snvs_name = default_SNVs(object),
  method = "BMix",
  upper_VAF_limit = 0.75,
  verbose = TRUE
)

fit_subclones_bmix(
  object,
  N = 1:3,
  powerlaw_model_name = active_models(object),
  snvs_name = default_SNVs(object),
  upper_VAF_limit = 0.75,
  verbose = TRUE
)

fit_subclones_mclust(
  object,
  N = 1:3,
  powerlaw_model_name = active_models(object),
  snvs_name = default_SNVs(object),
  upper_VAF_limit = 0.75,
  verbose = TRUE
)
}
\arguments{
\item{object}{object}

\item{N}{numbers of clones to for models}

\item{powerlaw_model_name}{residual of which powerlaw model to use?
powerlaw_fixed/powerlaw_optim}

\item{snvs_name}{which snvs to to use?}

\item{method}{clustering method to use: BMix and mclust are currently supported.
While mclust is a 3-4 times faster method, the BMix method is more accurate
and usually fast enough.}

\item{upper_VAF_limit}{ignore variants with f higher than}

\item{verbose}{verbose?}
}
\description{
Fit clonal and subclonal components of the model to the residuals of the
power-law model
}
\section{Functions}{
\itemize{
\item \code{fit_subclones_bmix()}: Fit subclonal distributions to neutral model residuals using BMix

\item \code{fit_subclones_mclust()}: Fit subclonal distributions to neutral model residuals using mclust

}}
