% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevodata-construction.R
\name{get_CNVs_var_names}
\alias{get_CNVs_var_names}
\alias{get_CNVs_var_names.cevodata}
\title{Get names of CNV variables}
\usage{
get_CNVs_var_names(object, ...)

\method{get_CNVs_var_names}{cevodata}(object, which = default_CNVs(object), ...)
}
\arguments{
\item{object}{object}

\item{...}{other arguments}

\item{which}{CNV assay to use}
}
\description{
Get names of CNV variables
}
\section{Methods (by class)}{
\itemize{
\item \code{get_CNVs_var_names(cevodata)}: Get CNV variable names from cevodata object

}}
