% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_snvs.R
\name{get_frequency_measure_name}
\alias{get_frequency_measure_name}
\alias{get_frequency_measure_name.cevodata}
\alias{get_frequency_measure_name.cevo_snvs}
\title{Decide which mutation frequency measure to use}
\usage{
get_frequency_measure_name(object, ...)

\method{get_frequency_measure_name}{cevodata}(object, which_snvs = default_SNVs(object), ...)

\method{get_frequency_measure_name}{cevo_snvs}(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{other params (not used now)}

\item{which_snvs}{Which SNVs to use?}
}
\description{
Used during the intervalization of mutation frequencies. Uses CCF/2 if found
in the object, and VAF otherwise
}
