% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_sample_mutation_burden}
\alias{get_sample_mutation_burden}
\title{Get sample mutation burden}
\usage{
get_sample_mutation_burden(cd)
}
\arguments{
\item{cd}{cevodata object}
}
\description{
Get sample mutation burden
}
