% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolutionary_parameters.R
\name{get_selection_coefficients}
\alias{get_selection_coefficients}
\title{Get subclonal selection coefficients Williams}
\usage{
get_selection_coefficients(object, Nmax = 10^10)
}
\arguments{
\item{object}{cevodata object}

\item{Nmax}{Time when tumour is sampled (in tumour doublings)}
}
\description{
Get subclonal selection coefficients Williams
}
