% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_snvs.R
\name{intervalize_mutation_frequencies}
\alias{intervalize_mutation_frequencies}
\alias{intervalize_mutation_frequencies.cevodata}
\alias{intervalize_mutation_frequencies.cevo_snvs}
\title{Intervalize the mutation frequencies}
\usage{
intervalize_mutation_frequencies(object, ...)

\method{intervalize_mutation_frequencies}{cevodata}(
  object,
  which_snvs = default_SNVs(object),
  column = get_frequency_measure_name(object, which_snvs),
  bins = NULL,
  verbose = get_cevomod_verbosity(),
  ...
)

\method{intervalize_mutation_frequencies}{cevo_snvs}(
  object,
  column = get_frequency_measure_name(object),
  bins = NULL,
  verbose = get_cevomod_verbosity(),
  ...
)
}
\arguments{
\item{object}{\if{html}{\out{<cevodata>}} object}

\item{...}{Other args}

\item{which_snvs}{Which SNVs to use}

\item{column}{Which frequency measure column to intervalize? By default, uses
the CCF/2  if it is found in the SNV tibble, and VAF otherwise}

\item{bins}{Number of interval bins}

\item{verbose}{Verbose?}
}
\value{
\if{html}{\out{<cevodata>}} object
}
\description{
Intervalize the mutation frequencies for the subsequent analyses and plots.
Adds f_interval column to the SNV tibble
}
