% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNVs_plots.R
\name{layer_mutations}
\alias{layer_mutations}
\title{Adds mutations to SFS plots}
\usage{
layer_mutations(
  genes = NULL,
  drivers = NULL,
  data = NULL,
  color = "black",
  size = 3,
  shape = "impact",
  filter_fun = guess_filter_fun(shape),
  ...
)
}
\arguments{
\item{color}{color}

\item{size}{size}

\item{...}{other arguments passed to geom_point()}
}
\description{
Adds mutations to SFS plots
}
