% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutral_lm.R, R/powerlaw_neutral_tail.R
\name{layer_neutral_tail}
\alias{layer_neutral_tail}
\alias{layer_neutral_tail.cevodata}
\title{layer_neutral_tail}
\usage{
layer_neutral_tail(object, ...)

\method{layer_neutral_tail}{cevodata}(
  object,
  start = NA,
  end = 1,
  binwidth = 0.01,
  show.legend = FALSE,
  color = "black",
  size = 1,
  ...
)
}
\arguments{
\item{object}{model fits from fit_neutral_lm}

\item{...}{other arguments passed to geoms}

\item{start}{VAF value to start plotting model fit}

\item{end}{VAF value to end plotting model fit}

\item{binwidth}{binwidth, necessary to scale the model parameter so it can
fit the binarized SFS correctly}

\item{show.legend}{show legend?}

\item{color}{line color}

\item{size}{line size}
}
\description{
layer_neutral_tail
}
\section{Methods (by class)}{
\itemize{
\item \code{layer_neutral_tail(cevodata)}: Plot neutral exponential curve on SFS plot

}}
