% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_williams_Mf_1f.R
\name{Mf_1f}
\alias{Mf_1f}
\alias{calc_Mf_1f}
\alias{calc_Mf_1f.cevodata}
\alias{calc_Mf_1f.cevo_snvs}
\alias{plot_Mf_1f}
\alias{plot_Mf_1f.cevodata}
\alias{get_Mf_1f}
\title{William's M(f) ~ 1/f statistics}
\usage{
calc_Mf_1f(object, ...)

\method{calc_Mf_1f}{cevodata}(
  object,
  which_snvs = default_SNVs(object),
  column = mutation_frequencies_column_name(object, which_snvs),
  bins = 100,
  verbose = get_cevomod_verbosity(),
  ...
)

\method{calc_Mf_1f}{cevo_snvs}(
  object,
  column = if ("CCF" \%in\% names(object)) "CCF" else "VAF",
  bins = 100,
  verbose = get_cevomod_verbosity(),
  ...
)

plot_Mf_1f(object, ...)

\method{plot_Mf_1f}{cevodata}(
  object,
  bins = NULL,
  from = 0.1,
  to = 0.25,
  scale = TRUE,
  geom = "point",
  ...
)

get_Mf_1f(object, model_name = "Mf_1f", verbose = TRUE, ...)
}
\arguments{
\item{object}{SNVs tibble object}

\item{...}{Other arguments}

\item{which_snvs}{Which SNVs to use?}

\item{column}{VAF/CCF}

\item{bins}{Resolution of the cumulative tails calculation}

\item{verbose}{Verbose?}

\item{from}{min f to plot}

\item{to}{max f to plot}

\item{scale}{scale data?}

\item{geom}{ggplot geom to use, eg. geom_line()}

\item{model_name}{name of slot with Mf_1f statistics}
}
\value{
ggplot obj
}
\description{
Mf_1f columns description:
\itemize{
\item sample_id
\item f
\item n - number of mutations in the VAF interval
\item \code{M(f)} and \code{1/f} columns to plot William's statistics
}
}
\section{Functions}{
\itemize{
\item \code{calc_Mf_1f()}: Calculate Williams M(f) ~ 1/f and saves it to
cevodata$models$Mf_1f

\item \code{calc_Mf_1f(cevodata)}: Method for \if{html}{\out{<cevodata>}} object

\item \code{calc_Mf_1f(cevo_snvs)}: Method for <cevo_snvs> object

\item \code{plot_Mf_1f(cevodata)}: Plot M(f) ~ 1/f

\item \code{get_Mf_1f()}: Get Mf_1f

}}
\examples{
data("test_data")
test_data |>
  calc_Mf_1f()

test_data |>
  plot_Mf_1f()
}
