% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/williams_Mf_1f.R
\name{Mf_1f}
\alias{Mf_1f}
\alias{calc_Mf_1f}
\alias{calc_Mf_1f.cevodata}
\alias{calc_Mf_1f.tbl_df}
\alias{plot_Mf_1f}
\title{William's M(f) ~ 1/f statistics}
\usage{
calc_Mf_1f(object, ...)

\method{calc_Mf_1f}{cevodata}(object, digits = 2, ...)

\method{calc_Mf_1f}{tbl_df}(object, digits = 2, ...)

plot_Mf_1f(
  object,
  digits = 2,
  from = 0.1,
  to = 0.25,
  scale = TRUE,
  geom = "point",
  ...
)

plot_Mf_1f(
  object,
  digits = 2,
  from = 0.1,
  to = 0.25,
  scale = TRUE,
  geom = "point",
  ...
)
}
\arguments{
\item{object}{SNVs tibble object}

\item{...}{other arguments}

\item{digits}{resolution of the cumulative tails calculation}

\item{from}{min VAF to plot}

\item{to}{max VAF to plot}

\item{scale}{scale data?}

\item{geom}{ggplot geom to use, eg. geom_line()}
}
\value{
ggplot obj
}
\description{
Creates  cevodata$models$Mf_1f tbl with the groupping variables and:
\itemize{
\item VAF
\item n - number of mutations in the VAF interval
\item \code{M(f)} and \code{1/f} columns to plot William's statistics
}
}
\section{Functions}{
\itemize{
\item \code{calc_Mf_1f(cevodata)}: Calculate Williams M(f) ~ 1/f

\item \code{calc_Mf_1f(tbl_df)}: Calculate Williams M(f) ~ 1/f

\item \code{plot_Mf_1f()}: Plot M(f) ~ 1/f

}}
\examples{
data("tcga_brca_test")
tcga_brca_test |>
  calc_Mf_1f()

tcga_brca_test |>
  plot_Mf_1f()
}
