% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_snvs.R
\name{mutation_frequencies}
\alias{mutation_frequencies}
\alias{calc_mutation_frequencies}
\alias{calc_mutation_frequencies.cevodata}
\title{Calc mutation frequencies}
\usage{
calc_mutation_frequencies(object, ...)

\method{calc_mutation_frequencies}{cevodata}(object, method = "use_VAF", ...)
}
\arguments{
\item{object}{data object}

\item{...}{other arguments}

\item{method}{method: "use_VAF". Other methods will be implemented later}
}
\description{
Currently no method is implemented and this function only initializes
f column in active SNVs slot with values from VAF column
}
