% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_williams_Mf_1f.R
\name{plot.cevo_Mf_1f_tbl}
\alias{plot.cevo_Mf_1f_tbl}
\title{Plot M(f) ~ 1/f}
\usage{
\method{plot}{cevo_Mf_1f_tbl}(
  x,
  from = 0.1,
  to = 0.25,
  scale = TRUE,
  geom = "point",
  mapping = NULL,
  ...
)
}
\arguments{
\item{x}{tibble with calc_Mf_1f() results}

\item{from}{min f to plot}

\item{to}{max f to plot}

\item{scale}{scale data?}

\item{geom}{ggplot geom to use, eg. geom_line()}

\item{mapping}{mapping}

\item{...}{futher passed to geom_()}
}
\value{
ggplot obj
}
\description{
Plot M(f) ~ 1/f
}
\examples{
data("test_data")

test_data |>
  calc_Mf_1f() |>
  plot_Mf_1f()
}
