% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_SFS.R
\name{plot.cevo_SFS_tbl}
\alias{plot.cevo_SFS_tbl}
\title{Plot SFS}
\usage{
\method{plot}{cevo_SFS_tbl}(x, y_scaled = FALSE, ...)
}
\arguments{
\item{x}{tibble with calc_SFS() results}

\item{y_scaled}{logical}

\item{...}{futher passed to geom_()}
}
\value{
ggplot obj
}
\description{
Plot SFS
}
