% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFS.R
\name{plot.cevo_SFS_tbl}
\alias{plot.cevo_SFS_tbl}
\alias{plot_SFS}
\title{Plot SFS}
\usage{
\method{plot}{cevo_SFS_tbl}(x, y_scaled = FALSE, ...)

plot_SFS(dt, ...)
}
\arguments{
\item{x}{tibble with calc_SFS() results}

\item{y_scaled}{logical}

\item{...}{args passed to stat_SFS}

\item{dt}{tibble with VAF column. Might be groupped and contain the data for multiple samples}
}
\value{
ggplot obj
}
\description{
Plot SFS
}
\section{Functions}{
\itemize{
\item \code{plot_SFS()}: Plot SFS dsfdsfsdfsd

}}
\examples{
data("tcga_brca_test")
tcga_brca_test |>
  dplyr::group_by(sample_id) |>
  calc_SFS() |>
  plot()

tcga_brca_test |>
  plot_SFS() +
  geom_mutations(drivers = "BRCA")
}
