% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_powerlaw_tail_fixed.R
\name{plot_Mf_1f_fits}
\alias{plot_Mf_1f_fits}
\title{Plot M(f) ~ 1/f fits}
\usage{
plot_Mf_1f_fits(object, model_name = "powerlaw_fixed", ...)
}
\arguments{
\item{object}{cevodata object}

\item{...}{other params}
}
\description{
Plot M(f) ~ 1/f fits
}
