% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_models}
\alias{plot_models}
\title{Plot cevodata models}
\usage{
plot_models(object, ...)
}
\arguments{
\item{object}{cevodata object}

\item{...}{other arguments}

\item{neutral_tail}{TRUE,}

\item{subclones}{TRUE,}

\item{final_fit}{TRUE,}
}
\description{
Plot cevodata models
}
