% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_models.R
\name{plot_models}
\alias{plot_models}
\alias{plot_models.cevodata}
\title{Plot cevodata models}
\usage{
plot_models(object, ...)

\method{plot_models}{cevodata}(
  object,
  models_name = active_models(object),
  show_neutral_tail = TRUE,
  show_binomial_layer = FALSE,
  show_subclones = TRUE,
  show_final_fit = TRUE,
  params_neutral_tail = list(),
  params_bootstraps = list(),
  params_binomial = list(),
  params_subclones = list(),
  params_final_fit = list(),
  nrow = NULL,
  ncol = NULL,
  ...
)
}
\arguments{
\item{object}{\verb{<cevodata>} object}

\item{...}{Other arguments passed to plot_SFS()}

\item{models_name}{Name of the models to plot}

\item{show_neutral_tail}{\verb{<lgl>}}

\item{show_binomial_layer}{\verb{<lgl>}}

\item{show_subclones}{\verb{<lgl>}}

\item{show_final_fit}{\verb{<lgl>}}

\item{params_neutral_tail}{List of non-default params for neutral tail geom}

\item{params_bootstraps}{List of non-default params for neutral tail bootstraps
geom}

\item{params_binomial}{List of non-default params for binomial geom}

\item{params_subclones}{List of non-default params for subclones geom}

\item{params_final_fit}{List of non-default params for final fit geom}

\item{nrow}{Passed to facet_wrap}

\item{ncol}{Passed to facet_wrap}
}
\description{
Plot cevodata models
}
