% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclones_binomial.R
\name{plot_models}
\alias{plot_models}
\alias{plot_models.cevodata}
\title{Plot cevodata models}
\usage{
plot_models(object, ...)

\method{plot_models}{cevodata}(
  object,
  neutral_tail = TRUE,
  binomial_layer = FALSE,
  subclones = TRUE,
  final_fit = TRUE,
  ...
)
}
\arguments{
\item{object}{cevodata object}

\item{...}{other arguments}

\item{neutral_tail}{TRUE,}

\item{subclones}{TRUE,}

\item{final_fit}{TRUE,}
}
\description{
Plot cevodata models
}
