% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_williams_neutral_model.R
\name{plot_neutral_A_coefficients}
\alias{plot_neutral_A_coefficients}
\title{Plot 'a' coefficients for all fitted neutral models}
\usage{
plot_neutral_A_coefficients(object, ...)
}
\arguments{
\item{object}{cevodata object}

\item{...}{other parameters passed to geom}
}
\description{
Plot 'a' coefficients for all fitted neutral models
}
