% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{plot_residuals}
\alias{plot_residuals}
\alias{plot_sampling_rate}
\alias{plot_sampling_rate.cevodata}
\alias{plot_residuals_powerlaw_model}
\alias{plot_residuals_powerlaw_model.cevodata}
\alias{plot_residuals_full_model}
\alias{plot_residuals_full_model.cevodata}
\alias{plot_binomial_fits_vs_powerlaw_residuals_bars}
\title{Plot model residuals}
\usage{
plot_sampling_rate(object, ...)

\method{plot_sampling_rate}{cevodata}(object, mapping = NULL, geom = geom_point, ...)

plot_residuals_powerlaw_model(object, ...)

\method{plot_residuals_powerlaw_model}{cevodata}(
  object,
  models_name = active_models(object),
  mapping = NULL,
  geom = geom_point,
  fit_clones = TRUE,
  ...
)

plot_residuals_full_model(object, ...)

\method{plot_residuals_full_model}{cevodata}(object, mapping = NULL, geom = geom_point, ...)

plot_binomial_fits_vs_powerlaw_residuals_bars(
  object,
  models_name = active_models(object),
  mapping = NULL,
  geom = geom_bar,
  fit_clones = TRUE,
  ...
)
}
\arguments{
\item{object}{cevodata object}

\item{...}{other parameters}

\item{mapping}{mapping elements to overwrite the default mapping}

\item{geom}{geom to use}

\item{fit_clones}{plot clonal fits?}
}
\description{
Plot model residuals
}
\section{Functions}{
\itemize{
\item \code{plot_sampling_rate(cevodata)}: Plot sampling rate

\item \code{plot_residuals_powerlaw_model(cevodata)}: Plot residuals of the neutral model

\item \code{plot_residuals_full_model(cevodata)}: Plot residuals of the full model

\item \code{plot_binomial_fits_vs_powerlaw_residuals_bars()}: Plot binomial fits vs powerlaw residuals (barplot)

}}
