% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_plots.R
\name{plot_sequencing_depth}
\alias{plot_sequencing_depth}
\alias{plot_sequencing_depth.cevodata}
\title{Plot sequencing depth of cevodata}
\usage{
plot_sequencing_depth(object, ...)

\method{plot_sequencing_depth}{cevodata}(object, mapping = NULL, geom = geom_boxplot, ...)
}
\arguments{
\item{object}{cevodata}

\item{...}{params passed to geom}

\item{mapping}{customize mapping}

\item{geom}{change geom}
}
\description{
Plot sequencing depth of cevodata
}
\section{Methods (by class)}{
\itemize{
\item \code{plot_sequencing_depth(cevodata)}: Plot sequencing data of mutations in cevodata

}}
