% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_powerlaw_tail_fixed.R
\name{powerlaw_fixed_model}
\alias{powerlaw_fixed_model}
\alias{fit_powerlaw_tail_fixed}
\alias{fit_powerlaw_tail_fixed.cevodata}
\alias{layer_lm_fits}
\title{Fitting neutral models}
\usage{
fit_powerlaw_tail_fixed(object, ...)

\method{fit_powerlaw_tail_fixed}{cevodata}(
  object,
  rsq_treshold = 0.98,
  name = "powerlaw_fixed",
  pct_left = 0.05,
  pct_right = 0.95,
  verbose = TRUE,
  ...
)

layer_lm_fits(cd, model_name = "powerlaw_fixed", ...)
}
\arguments{
\item{object}{SNVs tibble object}

\item{...}{other params passed to geom_segment()}

\item{rsq_treshold}{R-squared tresholds to keep model as neutral}

\item{name}{name in the models' slot}

\item{pct_left}{drop pct of the lowerst frequency variants to improve fit}

\item{pct_right}{drop pct of the highest frequency variants to improve fit}

\item{verbose}{verbose?}

\item{cd}{cevodata}

\item{model_name}{modelname}
}
\description{
Creates  cevodata$models$neutral_model
}
\section{Functions}{
\itemize{
\item \code{layer_lm_fits()}: Add M(f) ~ 1/f models layer to M(f) ~ 1/f plot

}}
\examples{
data("tcga_brca_test")
snvs <- SNVs(tcga_brca_test) |>
  dplyr::filter(sample_id \%in\% c("TCGA-AC-A23H-01","TCGA-AN-A046-01"))

cd <- init_cevodata("Test") |>
  add_SNV_data(snvs) |>
  calc_Mf_1f() |>
  calc_SFS() |>
  fit_powerlaw_tail_fixed(rsq_treshold = 0.99)

plot(cd$models$Mf_1f, from = 0.05, to = 0.4, scale = FALSE) +
  layer_lm_fits(cd)
}
