% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevomod-persistent.R
\name{set_containers_dir}
\alias{set_containers_dir}
\title{Set the containers directory}
\usage{
set_containers_dir(dir)
}
\arguments{
\item{dir}{Path for containers}
}
\description{
Set the containers directory
}
