% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_SFS.R
\name{sfs}
\alias{sfs}
\alias{calc_SFS}
\alias{calc_SFS.cevodata}
\alias{calc_SFS.cevo_snvs}
\alias{plot_SFS}
\alias{plot_SFS.cevodata}
\alias{get_SFS}
\title{Site Frequency Spectra}
\usage{
calc_SFS(object, ...)

\method{calc_SFS}{cevodata}(
  object,
  which_snvs = default_SNVs(object),
  column = get_frequency_measure_name(object, which_snvs),
  bins = NULL,
  verbose = get_cevomod_verbosity(),
  ...
)

\method{calc_SFS}{cevo_snvs}(
  object,
  column = if ("CCF" \%in\% names(object)) "CCF" else "VAF",
  bins = NULL,
  verbose = get_cevomod_verbosity(),
  ...
)

plot_SFS(object, ...)

\method{plot_SFS}{cevodata}(object, mapping = NULL, ..., geom = "bar")

get_SFS(object, model_name = "SFS", verbose = TRUE, ...)
}
\arguments{
\item{object}{SNVs tibble object}

\item{...}{other arguments}

\item{which_snvs}{Which SNVs to use?}

\item{column}{VAF/CCF}

\item{bins}{Resolution of the cumulative tails calculation}

\item{verbose}{verbose?}

\item{mapping}{aes()}

\item{geom}{Geom}

\item{model_name}{name of slot with SFS statistics}
}
\description{
Site Frequency Spectra (or Variant Allele Frequency Spectra) are the main
statistic used by cevomod.
}
\section{Functions}{
\itemize{
\item \code{calc_SFS()}: Calculates spectra for all samples and saves and saves them
in cevodata$models$SFS tibble.

SFS columns description:
\itemize{
\item y number of mutations in the frequency interval
\item y_scaled with y values scaled to the range 0-1
}

\item \code{calc_SFS(cevodata)}: method for cevodata object

\item \code{calc_SFS(cevo_snvs)}: method for cevo_snvs object

\item \code{plot_SFS()}: Plot SFS

\item \code{plot_SFS(cevodata)}: Plot SFS

\item \code{get_SFS()}: Get SFS

}}
\examples{
data("test_data")

test_data |>
  calc_SFS() |>
  plot_SFS() +
  layer_mutations(test_data, drivers = "BRCA")
}
