% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFS.R
\name{sfs}
\alias{sfs}
\alias{calc_SFS}
\alias{calc_SFS.cevodata}
\alias{calc_SFS.tbl_df}
\alias{plot_SFS}
\alias{plot_SFS.cevodata}
\title{Site Frequency Spectra}
\usage{
calc_SFS(object, ...)

\method{calc_SFS}{cevodata}(object, digits = 2, ...)

\method{calc_SFS}{tbl_df}(object, digits = 2, ...)

plot_SFS(object, ...)

\method{plot_SFS}{cevodata}(object, ...)
}
\arguments{
\item{object}{SNVs tibble object}

\item{...}{other arguments}

\item{digits}{resolution of the cumulative tails calculation}
}
\description{
Creates  cevodata$models$SFS with the groupping variables and:
\itemize{
\item n columnt with the number of mutations in the VAF interval
\item x and y columns describing SFS
\item y_scaled with y values scaled to the range 0-1
}
}
\section{Functions}{
\itemize{
\item \code{calc_SFS(cevodata)}: Calculate SFS

\item \code{calc_SFS(tbl_df)}: Calculate SFS

\item \code{plot_SFS(cevodata)}: Plot SFS

}}
\examples{
data("tcga_brca_test")
tcga_brca_test |>
  calc_SFS()

tcga_brca_test |>
  plot_SFS()
}
