% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_SFS.R
\name{sfs}
\alias{sfs}
\alias{calc_SFS}
\alias{calc_SFS.cevodata}
\alias{calc_SFS.cevo_snvs}
\alias{plot_SFS}
\alias{plot_SFS.cevodata}
\alias{get_SFS}
\alias{get_SFS.cevodata}
\title{Site Frequency Spectra}
\usage{
calc_SFS(object, ...)

\method{calc_SFS}{cevodata}(object, bins = NULL, ...)

\method{calc_SFS}{cevo_snvs}(object, bins = NULL, ...)

plot_SFS(object, ...)

\method{plot_SFS}{cevodata}(object, mapping = NULL, ..., geom = "bar")

get_SFS(object, ...)

\method{get_SFS}{cevodata}(object, name = "SFS", verbose = TRUE, ...)
}
\arguments{
\item{object}{SNVs tibble object}

\item{...}{other arguments}

\item{bins}{resolution of the cumulative tails calculation}

\item{mapping}{aes()}

\item{geom}{geom}
}
\description{
Creates  cevodata$models$SFS with the groupping variables and:
\itemize{
\item n columnt with the number of mutations in the VAF interval
\item x and y columns describing SFS
\item y_scaled with y values scaled to the range 0-1
}
}
\section{Functions}{
\itemize{
\item \code{calc_SFS()}: Calculate SFS

\item \code{calc_SFS(cevodata)}: Calculate SFS

\item \code{calc_SFS(cevo_snvs)}: Calculate SFS

\item \code{plot_SFS()}: Plot SFS

\item \code{plot_SFS(cevodata)}: Plot SFS

\item \code{get_SFS()}: Get SFS

\item \code{get_SFS(cevodata)}: Get SFS

}}
\examples{
data("tcga_brca_test")

tcga_brca_test |>
  calc_SFS() |>
  plot_SFS() +
  layer_mutations(tcga_brca_test, drivers = "BRCA")
}
