% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_SFS.R
\name{sfs}
\alias{sfs}
\alias{calc_SFS}
\alias{plot_SFS}
\alias{plot_SFS.cevodata}
\title{Site Frequency Spectra}
\usage{
calc_SFS(object, ...)

plot_SFS(object, ...)

\method{plot_SFS}{cevodata}(object, ..., geom = "line", alpha = if (geom == "bar") 0.8 else 1)
}
\arguments{
\item{object}{SNVs tibble object}

\item{...}{other arguments}

\item{geom}{geom}

\item{alpha}{alpha}

\item{digits}{resolution of the cumulative tails calculation}
}
\description{
Creates  cevodata$models$SFS with the groupping variables and:
\itemize{
\item n columnt with the number of mutations in the VAF interval
\item x and y columns describing SFS
\item y_scaled with y values scaled to the range 0-1
}
}
\section{Functions}{
\itemize{
\item \code{calc_SFS()}: Calculate SFS

\item \code{plot_SFS()}: Plot SFS

\item \code{plot_SFS(cevodata)}: Plot SFS

}}
\examples{
data("tcga_brca_test")
tcga_brca_test |>
  calc_SFS()

tcga_brca_test |>
  plot_SFS() +
  layer_mutations(drivers = "BRCA")

SNVs(tcga_brca_test) |>
  dplyr::group_by(sample_id) |>
  calc_SFS() |>
  plot()
}
