% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_tung_durrett_model.R
\name{tung_durrett}
\alias{tung_durrett}
\alias{fit_tung_durrett_models}
\alias{fit_tung_durrett_models.cevodata}
\title{Fitting Tung Durrett models}
\usage{
fit_tung_durrett_models(object, ...)

\method{fit_tung_durrett_models}{cevodata}(
  object,
  name = "tung_durrett",
  pct_left = 0,
  pct_right = 0.98,
  av_filter = c(1/3, 1/3, 1/3),
  control = list(maxit = 1000, ndeps = c(0.1, 0.01)),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{cevodata}

\item{...}{other arguments passed to stats::optim()}

\item{name}{name in the models' slot}

\item{pct_left}{drop pct of the lowest frequency mutations; currently not used}

\item{pct_right}{drop pct of the highest frequency mutations; currently not used}

\item{av_filter}{average filter values to be applied to VAF}

\item{control}{control param of stats::optim()}

\item{verbose}{verbose?}
}
\description{
\code{fit_tung_durrett_models()} uses \code{stats::optim} to find optimal A and alpha
whch maximizes SFS area under the powerlaw curve (\emph{sampled} region of SFS and
the range of VAF values below the maximum SFS value does not count) and
minimizes negative error - where the curve is above the real SFS (\emph{sampled}
are does not count). Penalty for the negative error depends on the number of
points with the negative error value. Penalty value is the sum of error values
to the power of x, where x is the length of the vector of negative error values.
This allows the powerlaw curve to detach from the SFS for 1 or two bins, but
then the penalty rises extremely.
}
\examples{
data("tcga_brca_test")
cd <- tcga_brca_test |>
  dplyr::filter(sample_id \%in\% c("TCGA-AC-A23H-01","TCGA-AN-A046-01")) |>
  fit_tung_durrett_models()
}
